+++ node/index.ts
@@ -51,6 +51,7 @@ export default new Service<Clients, RecorderState, ParamsContext>({
     resolvers: {
       Query: resolvers.Query,
       Mutation: resolvers.Mutation,
+      Quote: resolvers.Quote,
     },
     schemaDirectives,
   },