+++ node/utils/Queue.ts
@@ -1,22 +1,60 @@
 import { QUOTE_DATA_ENTITY, QUOTE_FIELDS, SCHEMA_VERSION } from '../resolvers'
+import message from './message'
 
-const processItem = ({ ctx, item }: { ctx: Context; item: any }) => {
+const processItem = ({ ctx, item }: { ctx: Context; item: Quote }) => {
   const {
     clients: { masterdata },
     vtex: { logger },
   } = ctx
 
+  const { id, referenceName, organization, costCenter, updateHistory } = item
+
+  const status = 'expired'
+  const now = new Date()
+  const nowISO = now.toISOString()
+
+  const users = updateHistory.map((anUpdate) => anUpdate.email)
+  const uniqueUsers = [...new Set(users)]
+
+  const lastUpdate = nowISO
+  const update = {
+    date: nowISO,
+    email: 'noreply@vtexcommerce.com.br',
+    role: 'expiration-system',
+    status,
+    note: '',
+  }
+
+  updateHistory.push(update)
+
   masterdata
     .updateEntireDocument({
       dataEntity: QUOTE_DATA_ENTITY,
-      id: item.id,
-      fields: { ...item, status: 'expired' },
+      id,
+      fields: { ...item, lastUpdate, updateHistory, status },
     })
     .then(() => {
-      logger.info({ message: `## ${item.id} changed to expired` })
+      message(ctx)
+        .quoteUpdated({
+          users: uniqueUsers,
+          name: referenceName,
+          id,
+          organization,
+          costCenter,
+          lastUpdate: {
+            email: 'expiration-system',
+            note: '',
+            status: status.toUpperCase(),
+          },
+        })
+        .catch((error) => {
+          logger.error({ message: 'quoteExpired-emailError', error })
+        })
+
+      logger.info({ message: `quoteExpired`, quoteId: id })
     })
     .catch((error) => {
-      logger.error({ message: 'Update masterdata error', error })
+      logger.error({ message: 'quoteExpired-mdError', error })
     })
 }
 
@@ -26,11 +64,16 @@ export const processQueue = (ctx: Context) => {
     vtex: { logger },
   } = ctx
 
+  const now = new Date()
+
+  now.setDate(now.getDate() + 1)
+  const nowISO = now.toISOString()
+
   masterdata
     .searchDocuments({
       dataEntity: QUOTE_DATA_ENTITY,
       fields: QUOTE_FIELDS,
-      where: `status <> 'expired' AND expirationDate < ${new Date().toISOString()}`,
+      where: `status <> 'expired' AND expirationDate < ${nowISO}`,
       sort: 'creationDate ASC',
       schema: SCHEMA_VERSION,
       pagination: {
@@ -40,7 +83,10 @@ export const processQueue = (ctx: Context) => {
     })
     .then((data: any) => {
       if (Array.isArray(data)) {
-        logger.info({ message: `#### Items to be processed => ${data.length}` })
+        logger.info({
+          message: `expirationQueue-foundItems`,
+          itemsToBeProcessed: data.length,
+        })
 
         data.forEach((item) => {
           processItem({ ctx, item })
@@ -48,7 +94,7 @@ export const processQueue = (ctx: Context) => {
       }
     })
     .catch((error) => {
-      logger.error({ message: 'Queue ERROR', error })
+      logger.error({ message: 'expirationQueue-error', error })
       throw error
     })
 }